+++ docs/next/markdoc/nodes/fence.markdoc.ts
@@ -1,8 +1,10 @@
-import {nodes} from '@markdoc/markdoc';
-
 import {Code} from '../../components/markdoc/Code';
 
 export const fence = {
   render: Code,
-  attributes: nodes.fence.attributes,
+  attributes: {
+    language: {type: String},
+    dagimage: {type: String},
+    fullwidth: {type: Boolean},
+  },
 };
